{"schema":"decodex/commit/1","summary":"Extract frozen frame content filter planner","authority":"manual"}
