+++ native/macos-host/Sources/RsnapNativeHostKit/NSScreenDisplayID.swift
@@ -1,0 +1,8 @@
+import AppKit
+import CoreGraphics
+
+extension NSScreen {
+	var nativeDisplayID: CGDirectDisplayID? {
+		(deviceDescription[NSDeviceDescriptionKey("NSScreenNumber")] as? NSNumber)?.uint32Value
+	}
+}
